apps/backend/services/safety-moderation/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
PyJWT==2.10.1
